fix: fixed crash when occurs on standalone build with Vulkan API   (#259)
* wip

(cherry picked from commit 5ba2cb87177c3b2ab3f370a46e2f8af52fa436c7)

* wip

(cherry picked from commit b1595526a38f9bad9a961c6581974c50649d6c8e)

* fixed bug

(cherry picked from commit 12e428e7d9fddb095698e3f16c5772cef9a5619f)

* fixed bug

(cherry picked from commit 84ecf06167ad0ac5d25d79abfa70f2c444a1efc6)

* refactor

(cherry picked from commit b281cf2825de192aec9d99addd7ee24cb1d62316)

+++ Plugin~/WebRTCPlugin/Codec/NvCodec/NvEncoder.cpp
@@ -261,7 +261,7 @@ namespace webrtc
 
     void NvEncoder::SetRates(uint32_t bitRate, int64_t frameRate)
     {
-        m_frameRate = frameRate;
+        m_frameRate = static_cast<uint32_t>(frameRate);
         m_targetBitrate = bitRate;
         isIdrFrame = true;
     }
